requirements.txt
#requirements.txt

fastapi
uvicorn[standard]
python-multipart

Pillow

torch
torchaudio
torchvision
transformers
accelerate
numpy
safetensors